Initial ActiveRecord Adapter support (#1)
This implements some initial support for response `Shed` timeouts/deadlines
into ActiveRecord, creating a `Shed::ActiveRecord::Adapter` which wraps all DB
querying methods with a call to `Shed.ensure_time_left!` which will raise a
`Shed::Timeout` if there is no time left.

This implementation is simple and naive, it does not propagate the
`Shed.time_left_ms` value to the call itself. This means that it is possible
for these queries to overrun the deadline and for the request to continue
processeing until the next check of `Shed.ensure_time_left!` is made and aborts.

This is a good start though. Implementing proper timeout propagation for
individula database queries could be done in the future by setting
connection/session variables (such as `statement_timeout` in PostgreSQL and
`max_execution_time` in MySQL). Care needs to be taken to _reset_ these
variables after any databse call and there is a caveat that using this method
in conjuction with a connection pooler that does not isolate connections (i.e.
works at the statement or transaction level) would cause these session
variables to leak.