Allow setting a maximum timeout in server middlewares (#2)
Update Shed#with_timeout to only update the deadline based on the new
timeout _if_ it results in a deadline which is earlier than any
currently set deadline.

Add some new ruby and go middlewares which will set a default upper
bound timeout for requests. These can be used in conjuction with client
timeout propagation middlewares in order to set an upper bound on
timeouts.